[FIX] artifacts: register MCP tool and context skill in manifest
