+++ plugins/artifacts/manifest.json
@@ -8,6 +8,14 @@
   "provides": "artifacts",
   "public_api": true,
   "public_api_prefix": "/artifacts",
+  "virtual_tools": "virtual_tools.py",
+  "context_skills": [
+    {
+      "name": "artifacts-usage",
+      "title": "Using artifacts",
+      "file": "skills/using-artifacts.md"
+    }
+  ],
   "dependencies": {
     "required": [],
     "optional": []
